requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-core
python-dotenv
pypdf
chromadb
pydantic
langchain-text-splitters
sentence-transformers